refactoring checkpoint - switched alignment of main and tool buttons, changed the implementation of comboboxes, this version may not have all features enabled

+++ src/main/java/visualizer/MouseHandler.java
@@ -1,5 +1,6 @@
 package visualizer;
 
+import javax.swing.*;
 import javax.swing.undo.StateEdit;
 import java.awt.*;
 import java.awt.event.*;
@@ -22,35 +23,51 @@ public class MouseHandler extends MouseAdapter {
     @Override
     public void mousePressed(MouseEvent event) {
         source = event.getComponent();
-        pressed = event.getLocationOnScreen();
-        location = source.getLocation();
-        if (source instanceof Vertex) {
-            moveStateEdit = new StateEdit((Vertex) source);
+        if (source instanceof JButton) {
+            var button = (MenuButton) source;
+            new Thread(() -> {
+                for (float i = 1f; i >= 0.6f; i -= .1f) {
+                    button.alpha = i;
+                    button.repaint();
+                    try {
+                        Thread.sleep(1);
+                    } catch (Exception ignored) {
+                    }
+                }
+            }).start();
+        } else {
+            pressed = event.getLocationOnScreen();
+            location = source.getLocation();
+            if (source instanceof Vertex) {
+                moveStateEdit = new StateEdit((Vertex) source);
+            }
         }
     }
 
     @Override
     public void mouseClicked(MouseEvent event) {
-        StateEdit stateEdit = new StateEdit(service);
-        switch (service.getGraphMode()) {
-            case ADD_A_VERTEX:
-                service.createNewVertex(event);
-                break;
-            case ADD_AN_EDGE:
-                service.createNewEdge(event);
-                break;
-            case REMOVE_A_VERTEX:
-                service.removeVertex(event);
-                break;
-            case REMOVE_AN_EDGE:
-                service.removeEdge(event);
-                break;
-            case NONE:
-                if (service.getAlgorithmMode() != AlgMode.NONE) service.startAlgorithm(event);
-                break;
+        if (source instanceof Vertex || source instanceof Graph) {
+            StateEdit stateEdit = new StateEdit(service);
+            switch (service.getGraphMode()) {
+                case ADD_A_VERTEX:
+                    service.createNewVertex(event);
+                    break;
+                case ADD_AN_EDGE:
+                    service.createNewEdge(event);
+                    break;
+                case REMOVE_A_VERTEX:
+                    service.removeVertex(event);
+                    break;
+                case REMOVE_AN_EDGE:
+                    service.removeEdge(event);
+                    break;
+                case NONE:
+                    if (service.getAlgorithmMode() != AlgMode.NONE) service.startAlgorithm(event);
+                    break;
+            }
+            stateEdit.end();
+            service.getUndoableEditSupport().postEdit(stateEdit);
         }
-        stateEdit.end();
-        service.getUndoableEditSupport().postEdit(stateEdit);
     }
 
     @Override
@@ -71,9 +88,43 @@ public class MouseHandler extends MouseAdapter {
 
     @Override
     public void mouseReleased(MouseEvent event) {
-        if (!source.getLocation().equals(location)) {
+        if (source instanceof Vertex && !source.getLocation().equals(location)) {
             moveStateEdit.end();
             service.getUndoableEditSupport().postEdit(moveStateEdit);
+        }
+    }
+
+    @Override
+    public void mouseEntered(MouseEvent event) {
+        var comp = event.getComponent();
+        if (comp instanceof JButton) {
+            new Thread(() -> {
+                for (float i = .5f; i <= 1f; i += .03f) {
+                    ((MenuButton) comp).alpha = i;
+                    comp.repaint();
+                    try {
+                        Thread.sleep(10);
+                    } catch (Exception ignored) {
+                    }
+                }
+            }).start();
+        }
+    }
+
+    @Override
+    public void mouseExited(MouseEvent event) {
+        var comp = event.getComponent();
+        if (comp instanceof JButton) {
+            new Thread(() -> {
+                for (float i = 1f; i >= .5f; i -= .03f) {
+                    ((MenuButton) comp).alpha = i;
+                    comp.repaint();
+                    try {
+                        Thread.sleep(10);
+                    } catch (Exception ignored) {
+                    }
+                }
+            }).start();
         }
     }
 }
